Get rid of hard-coded fmt.Print
We now look up any functions defined in native packages that were loaded.
Probably the right long-term approach here is to either dramatically limit the
native packages to things like system calls, or to do code generation to
generate wrappers for loading native packages. Or maybe there's some
reflection-like trick for looking up a package value or function by name.

This required adding a compile context to the compile step, but that was
probably going to be necessary anyway.