Implement basic functions
Right now there isn't full support for multiple return values or other features.
Also, the implementation just uses regular function calls: each interpreted
function is wrapped in a native function, which is used everywhere. This makes
it possible to send functions to native code, but limits the control we have and
forces the interpreter to use a stack as deep as the interpreted program, which
isn't ideal.