Add basic support for structs
There's now the concept of an "interprted struct", which is a struct backed by a
map.

We do the really easy thing for now and don't even look at the struct
definition. Instead, we map assignments to storing the value as a map field, so
the struct is populated lazily as a map.

This won't work for interop, but the vision is to have both native structs and
interpreted structs, based on whether the struct lives in a native or
interpreted package.